Refactored common utilities interface
